Unreviewed, fix build breakage part 2 https://bugs.webkit.org/show_bug.cgi?id=246795
* Source/JavaScriptCore/wasm/WasmTypeDefinition.h:
(JSC::Wasm::toB3Type):
* Source/JavaScriptCore/wasm/generateWasmOpsHeader.py:

Canonical link: https://commits.webkit.org/255779@main
